Don't forget requested sv_maxclients; migrate cvar
If sv_maxclients is set to something less than or equal to the highest
client num in use, it won't be decreased all the way to the requested
value, only to the smallest that doesn't kick off any clients. The old
implementation forgets about the real user-requested value; instead make
it keep trying on every map change (by storing the requested num as the
latched value again).

Also migrate the cvar.